requirements.txt
anthropic==0.41.0
google-generativeai>=0.8.0
openai>=1.30.0
jira==3.8.0
ollama==0.4.4
langgraph==0.2.60
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-community==0.3.13
chromadb==0.6.3
gitpython==3.1.44
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
PyYAML==6.0.2
PyGithub==2.5.0
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-json-report==1.5.0
anyio==4.7.0
aiosqlite